+++ lib/real_time_chart.dart
@@ -1,3 +1,5 @@
+library real_time_chart;
+
 export 'src/chart_display.dart';
 export 'src/point.dart';
-export 'src/live_chart.dart';
+export 'src/real_time_chart.dart';