+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## 0.0.1
+## 0.0.1 (Initial Release)
 
-* TODO: Describe initial release.
+- Initial release of the real_time_chart package.
+- Added the RealTimeChart widget, which displays a real-time graph of data.
+- Supports positive and negative values.
+- Adapts height depending on max value.
+- Includes example usage with a stream of data.
+
+